refactor: move consts, types to separate file
