--- ui/apps/everest/src/components/cluster-form/resources/messages.ts
@@ -1,8 +1,0 @@
-export const Messages = {
-  sharding: {
-    invalid: 'Please fill in valid values for sharding',
-    min: (val: number) => `The value cannot be less than ${val}`,
-    max: (val: number) => `The value cannot be more than ${val}`,
-    odd: 'The value cannot be even',
-  },
-};
