+++ ui/apps/everest/src/components/cluster-form/resources/resources.tsx
@@ -26,45 +26,11 @@ import {
 import { DbWizardFormFields } from 'consts';
 import { DbType } from '@percona/types';
 import { getProxyUnitNamesFromDbType } from './utils';
-
-type Props = {
-  unit?: string;
-  unitPlural?: string;
-  options: string[];
-  resourceSizePerUnitInputName: string;
-  cpuInputName: string;
-  diskInputName?: string;
-  diskUnitInputName?: string;
-  memoryInputName: string;
-  numberOfUnitsInputName: string;
-  customNrOfUnitsInputName: string;
-  disableDiskInput?: boolean;
-  allowDiskInputUpdate?: boolean;
-  dbType: DbType;
-};
-
-type ResourceInputProps = {
-  unit: string;
-  unitPlural: string;
-  name: string;
-  label: string;
-  helperText: string;
-  endSuffix: string;
-  numberOfUnits: number;
-  disabled?: boolean;
-};
+import { ResourcesTogglesProps, ResourceInputProps } from './resources.types';
+import { Messages } from './messages';
 
 const humanizeResourceSizeMap = (type: ResourceSize): string =>
   humanizedResourceSizeMap[type];
-
-const resourcesCapacityExceeding = (
-  fieldName: string,
-  value: number | undefined,
-  units: string
-) =>
-  `Your specified ${fieldName} size exceeds the ${
-    value ? `${value.toFixed(2)} ${units}` : ''
-  } available. Enter a smaller value before continuing.`;
 
 const estimated = (value: string | number | undefined, units: string) =>
   value ? `Estimated available: ${value} ${units}` : '';
@@ -86,7 +52,11 @@ const checkResourceText = (
       fieldLabel === 'cpu' ? parsedNumber / 1000 : parsedNumber / 10 ** 9;
 
     if (exceedFlag) {
-      return resourcesCapacityExceeding(fieldLabel, processedValue, units);
+      return Messages.resourcesCapacityExceeding(
+        fieldLabel,
+        processedValue,
+        units
+      );
     }
     return estimated(processedValue.toFixed(2), units);
   }
@@ -163,7 +133,7 @@ const ResourcesToggles = ({
   customNrOfUnitsInputName,
   disableDiskInput,
   allowDiskInputUpdate,
-}: Props) => {
+}: ResourcesTogglesProps) => {
   const { isMobile, isDesktop } = useActiveBreakpoint();
   const { data: resourcesInfo, isFetching: resourcesInfoLoading } =
     useKubernetesClusterResourcesInfo();
@@ -263,7 +233,9 @@ const ResourcesToggles = ({
             </ToggleCard>
           ))}
           {dbType !== DbType.Mysql && (
-            <ToggleCard value={CUSTOM_NR_UNITS_INPUT_VALUE}>Cutsom</ToggleCard>
+            <ToggleCard value={CUSTOM_NR_UNITS_INPUT_VALUE}>
+              {Messages.customValue}
+            </ToggleCard>
           )}
         </ToggleButtonGroupInput>
         {numberOfUnits === CUSTOM_NR_UNITS_INPUT_VALUE && (
